Modified mode of operation
Added exception raising if a { or } is found inside a line, but not at
the start. STRING is now searched for on a single line, data found
before STRING will be disregarded. Added sorting by integer so the
resulting JSON output will have the strings sorted by number instead of
alphabetically.