Correct imports, add test that loads many maps from a directory
Changed imports to reflect the recent changes in naming for the
read filetype scripts

Added a test that loads WC3 maps from a directory and tries to create a
listfile for them. This allows us to more easily detect oddities in maps,
e.g. unsupported compression types or issues caused by decompression.